Refactor shutdown test app
